Feat: Update the lock screen.
