+++ app/src/main/res/layout/fragment_password.xml
@@ -14,6 +14,28 @@
     android:padding="16dp">
 
     <LinearLayout
+        android:id="@+id/header_title"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_margin="26dp"
+        android:orientation="vertical"
+        android:padding="16dp">
+
+        <TextView
+            android:id="@+id/lock"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:gravity="center_horizontal" />
+
+        <TextView
+            android:id="@+id/message"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:gravity="center_horizontal"
+            android:text="@string/unlock_settings" />
+    </LinearLayout>
+
+    <LinearLayout
         android:id="@+id/password_list"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
@@ -21,6 +43,7 @@
         android:padding="16dp">
 
         <EditText
+            style="@style/EditTextStyle"
             android:id="@+id/password_1"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
@@ -28,10 +51,10 @@
             android:gravity="center_vertical|center_horizontal"
             android:inputType="numberPassword"
             android:maxLength="1"
-            android:textSize="28sp"
             tools:ignore="Autofill,LabelFor,SpeakableTextPresentCheck" />
 
         <EditText
+            style="@style/EditTextStyle"
             android:id="@+id/password_2"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
@@ -39,10 +62,10 @@
             android:gravity="center_vertical|center_horizontal"
             android:inputType="numberPassword"
             android:maxLength="1"
-            android:textSize="28sp"
             tools:ignore="Autofill,LabelFor,SpeakableTextPresentCheck" />
 
         <EditText
+            style="@style/EditTextStyle"
             android:id="@+id/password_3"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
@@ -50,10 +73,10 @@
             android:gravity="center_vertical|center_horizontal"
             android:inputType="numberPassword"
             android:maxLength="1"
-            android:textSize="28sp"
             tools:ignore="Autofill,LabelFor,SpeakableTextPresentCheck" />
 
         <EditText
+            style="@style/EditTextStyle"
             android:id="@+id/password_4"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
@@ -61,16 +84,7 @@
             android:gravity="center_vertical|center_horizontal"
             android:inputType="numberPassword"
             android:maxLength="1"
-            android:textSize="28sp"
             tools:ignore="Autofill,LabelFor,SpeakableTextPresentCheck" />
     </LinearLayout>
-
-    <Button
-        android:id="@+id/submit_button"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        style="@style/ButtonStyle"
-        android:text="@string/unlock_settings"
-        tools:ignore="VisualLintButtonSize" />
 
 </LinearLayout>
